feat: generalize and collect all MatchingAlgorithm classes
